Moved NBA settings to settings file
